+++ src/Helper/AktualisiereVereine.php
@@ -13,7 +13,42 @@ class AktualisiereVereine
 
 	public static function fromVereinsliste($result)
 	{
-		self::fromSpielerliste($result);
+		//echo '<pre>';
+		//print_r($result);
+		//echo '</pre>';
+
+		if($result->union)
+		{
+			// Verein in lokaler Datenbank suchen
+			$objVerein = \Database::getInstance()->prepare("SELECT * FROM tl_dwz_ver WHERE zpsver = ?")
+			                                     ->execute($result->union->vkz);
+
+			// Datensatz füllen
+			$set = array
+			(
+				'tstamp'     => time(),
+				'zpsver'     => $result->union->vkz ? $result->union->vkz : '',
+				'name'       => $result->union->name,
+				'elobase'    => '',
+				'published'  => 1,
+			);
+
+			// Datensatz eintragen
+			if($objVerein->numRows)
+			{
+				// Vereinsname aktualisieren (einziges Feld neben vkz, was die Spielersuche liefert)
+				$objUpdate = \Database::getInstance()->prepare("UPDATE tl_dwz_ver %s WHERE id = ?")
+				                                     ->set($set)
+				                                     ->execute($objVerein->id);
+			}
+			else
+			{
+				// Verein in lokaler Datenbank nicht gefunden, deshalb neu anlegen
+				$objInsert = \Database::getInstance()->prepare("INSERT INTO tl_dwz_ver %s")
+				                                     ->set($set)
+				                                     ->execute();
+			}
+		}
 	}
 
 	public static function fromSpielerliste($result)